feat(ci): add CI
Also disables `nix flake check` support. This is because we'd have to
mess about with including the necessary Python dependencies in the
`nix flake check` invocation so that Pyright doesn't error.

What we can do instead to run checks in CI is `nix develop --command pre-commit run
-a`. This uses the devShell which already has the Python dependencies
from poetry2nix.